Implementation Tested and changed
I fixed your broken code!! Just kidding, your implementation was spot on
except when it got to a day before the due date it didn't have a
reminder. I also fixed some of the color, the reminders turned to white
on some weeks. Weird. So I went ahead and added colors to the weekly
reminders and daily reminders. So it is easy to tell the difference. I
would say this user story has been completed.