Create Class Task
This isn't complete yet, what is done is the framework for what we will
be trying to implement, I still don't know how memoranda captures the
system time. But now we have a class to allocate values and functions
to. First step. I commented areas in the event code that should be
useful.