Calc Time Implementation
So thhis is the function to calculate time between the two dates. I
think I may have simplified a lot of our problems by doing it this way.
We no longer have to rely on the calendar to capture the time, as that
would skew our data. This function still needs reviewing. Need help on
what to do next with the vectors.